Log replication complete
No longer treating heartbeats any differently than empty Append Entries
RPCs.  We are now updating the logs of followers with heartbeats by
sending logs that are missing to bring them up to date as fast
as possible, instead of waiting for another operation to be executed
for us to update the followers' potentially outdated log.